+++ hr_timesheet_task_required/views/hr_timesheet_sheet.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <record model="ir.ui.view" id="hr_timesheet_sheet_form">
+        <field name="name">hr.timesheet.sheet.form (in hr_timesheet_task_required)</field>
+        <field name="model">hr_timesheet_sheet.sheet</field>
+        <field name="inherit_id" ref="hr_timesheet_task.hr_timesheet_sheet_form"/>
+        <field name="arch" type="xml">
+
+            <xpath expr="//field[@name='timesheet_ids']/tree/field[@name='task_id']" position="attributes">
+                <attribute name="required">1</attribute>
+            </xpath>
+
+        </field>
+    </record>
+
+</odoo>
