--- hr_timesheet_task_required/views/hr_timesheet_task_required.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-        <template id="assets_backend" name="hr_timesheet_task_required" inherit_id="web.assets_backend">
-            <xpath expr="." position="inside">
-                <script type="text/javascript" src="/hr_timesheet_task_required/static/src/js/hr_timesheet_widget.js"></script>
-            </xpath>
-        </template>
-    </data>
-</openerp>
