--- hr_timesheet_task_required/views/hr_timesheet_view.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-        <record model="ir.ui.view" id="hr_timesheet_sheet_form_with_activity">
-            <field name="name">hr.timesheet.sheet.form (hr_timesheet_task_required)</field>
-            <field name="model">hr_timesheet_sheet.sheet</field>
-            <field name="inherit_id" ref="hr_timesheet_task.hr_timesheet_sheet_form_with_activity"/>
-            <field name="arch" type="xml">
-                <field name="task_id" position="attributes">
-                    <attribute name="required">1</attribute>
-                </field>
-            </field>
-        </record>
-
-        <record id="hr_timesheet_line_form_task_required" model="ir.ui.view">
-            <field name="name">hr.analytic.timesheet.form (hr_timesheet_task_required)</field>
-            <field name="model">hr.analytic.timesheet</field>
-            <field name="inherit_id" ref="hr_timesheet_task.hr_timesheet_line_form"/>
-            <field name="arch" type="xml">
-                <field name="task_id" position="attributes">
-                    <attribute name="required">1</attribute>
-                </field>
-            </field>
-        </record>
-
-        <record id="hr_timesheet_line_tree_task_required" model="ir.ui.view">
-            <field name="name">hr.analytic.timesheet.tree  (hr_timesheet_task_required)</field>
-            <field name="model">hr.analytic.timesheet</field>
-            <field name="inherit_id" ref="hr_timesheet_task.hr_timesheet_line_tree"/>
-            <field name="arch" type="xml">
-                <field name="task_id" position="attributes">
-                    <attribute name="required">1</attribute>
-                </field>
-            </field>
-        </record>
-    </data>
-</openerp>
